new class: notations
